1. get_regions_skus_sps 기능을 기존 파일에서 분리나옴. 2. files_sps 폴더에  config_sps.ini 파일 추가하여 파일 경로오 subscription_id 값 저장.
