+++ collector/spot-dataset/azure/lambda/current_collector/files_sps/config_sps.ini
@@ -1,0 +1,12 @@
+[Config]
+SUBSCRIPTION_ID = e9c8c784-8c09-45ac-88da-3b2c206c22f6
+# CLI 위치
+AZ_CLI_PATH = C:\\Program Files\\Microsoft SDKs\\Azure\\CLI2\\wbin\\az.cmd
+#1차로 수집된 리전과 리전 내의 VM들
+REGION_VM_SKU_SOURCE_FILENAME = ./files_sps/region_vm_sku_source.json
+# 리전과 VM 데이터를 유니크하게 가공하고 이용 가능한 형식으로 변환
+REGION_VM_SKU_PROCESSED_FILENAME = ./files_sps/region_vm_sku_processed.json
+# get_spot_placement_recommendation 실행 시 지원되지 않는 리전을 별도 파일에 저장하여 참조
+INVALID_REGIONS_FILENAME = ./files_sps/invalid_regions.json
+# spot_placement_recommendation 결과
+RESULTS_FOR_RECOMMENDATION_FILENAME = ./files_sps/merged_results.json
